requirements.txt
discord.py
pynacl
youtube-dl